Rewrote notes to work with user IDs
